Collections/BaseDictionaries.cs: Collections/BaseLists.cs: Collections/BaseSets.cs: Collections/CircularList.cs: Collections/Stack.cs: * Made the public GetEnumerator() return the actual type rather than IEnumerable   in the same way as the built-in collections do (for performance)
Collections/BloomFilter.cs:
* Replaced (uint)1 with 1u

Collections/PriorityQueue.cs:
* Tweaked an exception message

IO/ArrayIO.cs:
* Made Read* and Write* functions use optimized versions when the endianness of
  the function matched the endianness of the architecture
* Created ReadLESingle, ReadBESingle, ReadLEDouble, and ReadBEDouble to replace
  ReadSingle and ReadDouble, since apparently the endianness of IEEE 754 numbers
  is variable. Ditto for the Write methods.

IO/BinaryReaderWriter.cs:
* Made float and double endianness variable
* Inlined a few common operations

IO/Streams.cs:
* Renamed ArrayBuffer.GetContiguousArray to .GetZeroOffsetArray

Mathematics/BigInteger.cs:
Mathematics/Tests/BigInteger.cs:
+ Added Integer, an arbitrary-precision integer class

Mathematics/FloatingPoint.cs:
* Separated the internal arbitrary-precision integer class into a fully fledged public type
* Refactored the number printing and parsing code out of this class

Mathematics/Functions.cs:
Mathematics/Optimization.cs:
Mathematics/Polynomials.cs:
Mathematics/RootFinding.cs:
Mathematics/Vector.cs:
* Renamed ArrayUtility.SmallCopy to .FastCopy

Mathematics/Geometry.cs:
* Added MathConst.Sqrt2 and .HalfSqrt2

Mathematics/Matrices.cs:
* Added some #region lines

Mathematics/NumberFormatting.cs:
* Refactored number printing and parsing out of FP107 to be shared with Integer

Mathematics/Random.cs:
+ Added RandomNumberGenerator.NextExponential and .NextNormal, which are pretty fast
  methods for generating numbers from the exponential and normal distributions
* Made methods that take seeds use the 'params' keyword
* Used KISSRNG rather than XorShift128RNG to fill out the MWC256RNG seed

Mathematics/Tests/FloatingPoint.cs:
* Added tests to reproduce some problems exposed by refactoring code out of FP107

Utilities/ArrayUtility.cs:
Utilities/ArrayUtility.tt:
* Renamed ArrayUtility.SmallCopy to .FastCopy
* Made ArrayUtility.FastCopy use a for loop for small arrays, to accept arrays of
  reference types, and to eliminate the Array.Copy path in most cases

Utilities/Arrays.cs:
Utilities/Tests/ArrayTests.cs:
* Made the public GetEnumerator() return the actual type rather than IEnumerable
  in the same way as the built-in collections do (for performance)
* Renamed ArrayBuffer.GetContiguousArray to .GetZeroOffsetArray, and fixed a bug
  that could cause the array to be modified even if PinData was true

Utilities/BinaryUtility.cs:
Utilities/Tests/BinaryTests.cs:
+ Added BinaryUtility.CountBits, to implement the population count operation
* Added some explanatory comments
+ Added IsHex(char), ParseHex(char), ToHexChar(byte), and TryParseHex(char, out byte)
+ Added a ToHex(byte[], int, int) override
* Fixed a bug in Hash that could cause reading past the end of an array in some cases

Utilities/Linq.cs:
+ Added IEnumerable<T>.AsCollection and .AsList extensions
+ Added IEnumerable<T>.SelectIndexValuePairs, .SelectValueIndexPairs, ToIndexValueDictionary,
  and ToValueIndexDictionary extensions to perform the somewhat-common operation of
  selecting a collection of values and their indexes
+ Added an IEnumerable<T>.ToArray extension that effeciently selects elements into an array

Utilities/Utility.cs:
+ Added a method Utility.EnlargeArray override that doesn't leave any space for future growth

+++ Collections/BloomFilter.cs
@@ -143,7 +143,7 @@ public class BloomFilter<T>
     {
       // we'll get the value from each hash function and use it as an index into the bit array, setting the corresponding bits
       uint hash = (uint)hashProvider.GetHashCode(hashFunction, item) % bitCount;
-      bits[hash >> 5] |= (uint)1 << (int)(hash & 31);
+      bits[hash >> 5] |= 1u << (int)(hash & 31);
     }
   }
 
@@ -164,7 +164,7 @@ public class BloomFilter<T>
       // we'll get the value from each hash function and use it as an index into the bit array. the item has not been added if
       // any bit is zero, since Add() would have set all of those bits
       uint hash = (uint)hashProvider.GetHashCode(hashFunction, item) % bitCount;
-      if((bits[hash >> 5] & ((uint)1 << (int)(hash & 31))) == 0) return false;
+      if((bits[hash >> 5] & (1u << (int)(hash & 31))) == 0) return false;
     }
     return true; // all of the bits were set, so the item might have been added (or this may be a false positive)
   }
